test(l1): move P256VERIFY differential tests to the integration test crate
Relocate the aws-lc-rs vs pure-Rust p256 differential tests from an inline
#[cfg(test)] module in p256_awslc.rs to test/tests/crypto/p256_tests.rs.
They now exercise the public API (NativeCrypto + a no-override Crypto impl
as the reference), so no private items are needed.
