+++ test/tests/crypto/mod.rs
@@ -1,2 +1,3 @@
 mod blake2f_tests;
 mod keccak_tests;
+mod p256_tests;
